+++ Quadtastic/Quadtastic.lua
@@ -258,6 +258,15 @@ Quadtastic.draw = function(app, state, gui_state)
         Tooltip.draw(gui_state, S.tooltips.strip_tool)
         Layout.next(gui_state, "|")
 
+        if Button.draw(gui_state, nil, nil, nil, nil, nil,
+                       gui_state.style.quads.tools.wand,
+                       {pressed = state.toolstate.type == "wand"})
+        then
+          state.toolstate = {type = "wand"}
+        end
+        Tooltip.draw(gui_state, S.tooltips.wand_tool)
+        Layout.next(gui_state, "|")
+
       end Layout.finish(gui_state, "|")
 
       Layout.next(gui_state, "-")